search lib
made to a separate branch incase i messed something up really bad while fumbling around with the classes. Also brought over a string checker i made for the state project. It might be a tad confusing, there are a lot of comments, i needed them when i was making it to track what it was doing and keep things straight, we can delete these comments later if we want.
searchlib will return an EnumerableRowCollection which we should be able to do:
____DataGridView.DataSource = WhatToSearch(string league, string searchWhat, string searchVal) // searched for searchVal within a string
or
____DataGridView.DataSource = WhatToSearch(string league, string searchWhat, string minVal, string maxVal) // searches between the min and max.

we should only need to pass in user input as strings and then the search will confirm if the values are valid. It then searches or displays a warning message to the user